Remove handlers.numbers
